requirements.txt
# apiflask
APIFlask==1.1.3
apispec==5.2.2
click==8.1.3
Flask==2.1.3
Flask-HTTPAuth==4.7.0
flask-marshmallow==0.14.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.18.0
packaging==21.3
pyparsing==3.0.9
six==1.16.0
webargs==8.2.0
Werkzeug==2.2.2
zipp==3.8.1

# python-dotenv
python-dotenv==0.21.0

# requests
certifi==2022.9.14
charset-normalizer==2.1.1
idna==3.4
requests==2.28.1
urllib3==1.26.12

# geopy
geographiclib==1.52
geopy==2.2.0
